修复 计算大小bug

+++ Runtime/QByte.cs
@@ -303,7 +303,7 @@ namespace QTool.ByteExtends
 
         public static string ComputeScale(this int byteLength)
         {
-            return ComputeScale(byteLength);
+            return ComputeScale((long)byteLength);
         }
 
         public static string ComputeScale(this long byteLength)
